refactor pm quality review action preview router
